Step2 - 인수 테스트 리팩토링 (#83)
* Upload 1step 3 test cases

* Upload step1 test cases

* Upload Step 2 - Acceptance Test Refactoring

* Delete list of lines assert at LineAcceptanceTest.java

* Add custom exception when update line

* Add NoSuchLineException at LineController.java

* Add lineCreatedHelper method

Fix
- Pascal Convention to java convention

* Fix wrong java convention

* Refactor station acceptance test

+++ src/main/java/nextstep/subway/station/application/StationService.java
@@ -1,13 +1,18 @@
 package nextstep.subway.station.application;
 
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.exception.DuplicateLineException;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationRepository;
 import nextstep.subway.station.dto.StationRequest;
 import nextstep.subway.station.dto.StationResponse;
+import nextstep.subway.station.exception.DuplicateStationException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -20,20 +25,26 @@ public class StationService {
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
-        Station persistStation = stationRepository.save(stationRequest.toStation());
-        return StationResponse.of(persistStation);
+        try {
+            Station persistStation = stationRepository.save(stationRequest.toStation());
+            return StationResponse.of(persistStation);
+        } catch (DataIntegrityViolationException e){
+            throw new DuplicateStationException("Already registered Station");
+        }
     }
 
     @Transactional(readOnly = true)
     public List<StationResponse> findAllStations() {
         List<Station> stations = stationRepository.findAll();
-
         return stations.stream()
                 .map(station -> StationResponse.of(station))
                 .collect(Collectors.toList());
     }
 
     public void deleteStationById(Long id) {
-        stationRepository.deleteById(id);
+        Optional< Station > optionalStation = stationRepository.findById(id);
+        if(optionalStation.isPresent()) {
+            stationRepository.deleteById(id);
+        }
     }
 }
